Started on "handle_400_response" function.
Filled out the first "handle_400_response" and "parse_400_response"
functions. The "handle" function returns a dictionary that maps all
the errors returned in the JSON response to a known error type,
given in the ClassroomPostErrors enum. The plan is to have a similar
enum and function in each module for handling the 400 responses.