Added check for over-sized batch.
Added a class variable `max_batch_size` that specifies the maximum number of objects
that can be submitted to a single POST or PUT call. To support this, a new
ApexMaxPostSizeException is thrown in the `post_batch` if the length of `objects`
is greater than the `max_batch_size` variable of a given class.