Added support for batch tokens.
Attempting to POST and PUT a large enough batch will sometimes return a batch status token
instead of an immediate response. This commit added support for these instances. The
ApexSynchronizer class will wait a certain amount of time, defined by
`MAX_BATCH_WAIT_TIME` in the `utils` module (set at 90 seconds in this commit), checking
every second to see if the batch is complete. When the batch is done, the POST and PUT
will return the response. If after the specified amount of time the batch is still
processing, an `ApexBatchTimeoutError` is thrown.