Changes debug output in classroom enrollment sync.
Instead of always outputting the percent of errors from an enrollment
PUT call that are "already exist"/"already enrolled" errors, it only
outputs this percentage when enrollment records have changed, i.e.
when not all of the errors are already enrolled errors. Ideally,
these errors should be detected before making the PUT call at all,
but this will suffice for now.