Updates to Apex are now recorded in the sync context.
Because interfacing with Apex is a lengthy process, the
synchronizer gathers all necessary information during initialization.
Before, updates to Apex were not recorded in the local context,
meaning that if a student was added during one part of the sync,
this would not be seen until Apex was contacted again, i.e. the
next sync. Now the local `ApexEnrollment` object is updated
whenever a PUT, POST, or DELETE call is made to the server.

This should fix bugs where students were added during the
`sync_roster` operation but not enrolled because the information
in the ApexEnrollment object was not current.