Fixed __del__ namespace bug.
In newer versions of Python, the global reference to
Python's build-in `open` function is at risk of having
been removed from the namespace during object
deconstruction, i.e. during the call to __del__. As such
automatically saving operations to a JSON file on
deconstruction is bug-prone. Changed the __del__ function
to a `save` function, that is called in `main.py`.