Fixed bug when only roster change is a transfer
During a dry run of the `sync_roster` method, if no changes
were found before a call to the `_find_conflicts` helper
method, the `sync_roster` sub-dictionary of the `_operations`
attribute object was accessed, leading to a KeyError because
that dictionary was not created yet. Simply checks for that
key, and if it is not found, adds it to the attribute.